requirements.txt
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.1
Flask==2.2.3
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.2
python-dotenv==1.0.0
tomli==2.0.1
Werkzeug==2.2.3
zipp==3.15.0
